Version 3 (#28)
* Kicking off version 3 development - #27

* Version 3 complete for Cosmos & TableStorage #27

* Cleaning up collection -> container

* Fixing InMemory + dependencies

* Position changed to Version but with back-compat #27

* Tests refactor to test back-compat WIP #27

* Back compat now fully tested for CosmosDb #27

* Back compat now fully tested for TableStorage #27

* Show error version in exception message #26

* Ready for v3 release

+++ tests/CosmoStore.TableStorage.Tests/Program.fs
@@ -1,23 +1,24 @@
 module CosmoStore.TableStorage.Tests.Program
 
 open System
+open System.Diagnostics
 open Expecto
 open Expecto.Logging
 open CosmoStore.Tests
 open CosmoStore.TableStorage
 open Microsoft.WindowsAzure.Storage
 
-let private tableName = "CosmoStoreTests"
+let private v3config = Configuration.CreateDefaultForLocalEmulator() |> fun cfg -> { cfg with TableName = "CosmoStoreTests" }
+let private v2config = { v3config with TableName = "CosmoStoreTestsV2" }
 
-let private conf = Configuration.CreateDefaultForLocalEmulator() |> fun cfg -> { cfg with TableName = tableName }
-
-let getCleanEventStore() =
+let getEventStore cleanup conf =
     let account = CloudStorageAccount.DevelopmentStorageAccount
     let client = account.CreateCloudTableClient()
-    let table = client.GetTableReference(tableName)
-    try
-        table.DeleteIfExistsAsync() |> Async.AwaitTask |> Async.RunSynchronously |> ignore
-    with _ -> ()
+    let table = client.GetTableReference(conf.TableName)
+    if cleanup then
+        try
+            table.DeleteIfExistsAsync() |> Async.AwaitTask |> Async.RunSynchronously |> ignore
+        with _ -> ()
     conf |> EventStore.getEventStore
  
 let testConfig = 
@@ -25,10 +26,12 @@ let testConfig =
         parallelWorkers = 2
         verbosity = LogLevel.Debug }
 
-let cfg = Domain.defaultTestConfiguration getCleanEventStore
-
 [<EntryPoint>]
 let main _ = 
-    (cfg, "Table Storage") 
-    |> AllTests.getTests 
-    |> runTests testConfig
+    // run v3 tests
+    AllTests.getTests "Table Storage - V3" Generator.defaultGenerator (getEventStore true v3config) |> runTests testConfig |> ignore
+    // write "old" events
+    Process.Start("dotnet", "run -p tests/CosmoStore.TableStorage.Tests.V2").WaitForExit()
+    // run v2 tests
+    AllTests.getTests "Table Storage - V2" Generator.defaultGenerator (getEventStore false v2config) |> runTests testConfig |> ignore
+    0
